+++ src/Classifiers/GaussianNB.php
@@ -34,7 +34,7 @@ class GaussianNB implements Online, Probabilistic, Persistable
 {
     const TWO_PI = 2. * M_PI;
 
-    const LOG_EPSILON = -8;
+    const LOG_EPSILON = -18.420680744;
 
     /**
      * The class prior probabilities.