+++ src/CrossValidation/Reports/ConfusionMatrix.php
@@ -2,6 +2,7 @@
 
 namespace Rubix\ML\CrossValidation\Reports;
 
+use Rubix\ML\Estimator;
 use InvalidArgumentException;
 
 /**
@@ -44,6 +45,19 @@ class ConfusionMatrix implements Report
         }
 
         $this->classes = $classes;
+    }
+
+    /**
+     * The estimator types that this report is compatible with.
+     * 
+     * @return int[]
+     */
+    public function compatibility() : array
+    {
+        return [
+            Estimator::CLASSIFIER,
+            Estimator::DETECTOR,
+        ];
     }
 
     /**