+++ src/CrossValidation/Reports/Report.php
@@ -7,6 +7,13 @@ interface Report
     const EPSILON = 1e-8;
 
     /**
+     * The estimator types that this report is compatible with.
+     * 
+     * @return int[]
+     */
+    public function compatibility() : array;
+
+    /**
      * Generate the report.
      *
      * @param  array  $predictions