+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     - Added stack method to dataset API
     - Changed merge method on dataset to append and prepend
     - Added compatibility method to validation metrics
+    - Added estimator compatibility to reports
 
 - 0.0.8-alpha
     - Added Model Orchestra meta estimator